Add an option to specify whitelisted tiers in CRAB submission
If the user wants to submit the jobs to a selection of tiers, then the list of tiers must be given as a single comma-separated string, e.g.

-w T2_EE_Estonia,T2_IT_Bari

This information is propagated to the crab cfg file via shell environment variable named $WHITELIST. If no whitelist is provided, then the job submission proceeds as in the default case.