Script for producing PU distributions in data
The main script -- compPileUp.py -- calls CMSSW's pileupCalc.py which unfortunately contains a syntax error in 10_2_10. For this reason I also added pileupCalc.py taken from
https://github.com/cms-sw/cmssw/blob/3dd24ec2c6d8a33ec5b4a5d3163f8d0bec6bee1c/RecoLuminosity/LumiDB/scripts/pileupCalc.py
and made the path to pileupCalc.py configurable in compPileUp.py (the default setting is to use the script that's in this repository). Looks like the syntax error is present in all CMSSW 10X versions prior to 103X.