Automatically detect tH samples while submitting CRAB jobs
The crux of this problem is generating two different nano cfg files -- one for the tH samples and the other for any other sample. So, the solution is to automatically generate additional cfg file for tH sample only if the current submission is meant for MC samples. Otherwise, the program just bails out (this prevents the user from submitting data jobs with tH cfg or generating tH cfg files for sync jobs). In order to make things easier from the backend side, I've disabled the command line parameter which lets the user to specify the nano cfg file. In practice, this use case has become obsolete anyways.