List of datasets having a file block w/ more than 100k lumis
Some datasets have a file block that contains more than 100k luminosity blocks. This causes problems w/ CRAB submission b/c the CRAB servers decline jobs that process such datasets when submitting w/ the EventAwareLumiBased splitting option (it would generate a huge number of jobs which causes CRAB servers to crash). The current limit on the number of luminosity block a dataset can have in a file block is currently at 100k. If the dataset contains more than 100k lumis, FileBased splitting option must be used.
However, this causes another problem: the files in those datasets contain too many events per file and the jobs wouldn't finish in the recommended time frame of 8h to 16h. The solution is to generate multiple nano cfg files that process some part of the file. The way a given file is split by multiple jobs depend on the number of events in the file and on the parameter that specifies how many events the job should process. Therefore, we need to know the maximum number of events in the files of the dataset, in order to generate enough nano cfg files in advance before the actual job submission. For instance, if a dataset has the file block issue and it has a file containing 140k events, we would need to generate 3 nano cfg files, assuming that each job processes 50k events: the 1st and 2nd job processes 50k events and the 3rd job processes 40k events.
File datasets_fileblock_err.txt is a preliminary list of datasets that have the file block issue. Each line lists the dataset and maximum number of events found in all files of the dataset. This information will be later utilized in launchall.sh submission script. The file is generated w/ parse_lumiblock_errors.py on the output of count_lumis_datasets.sh.