Make the computation of jet charges and pulls O(100x) faster
The previous implementation had lots of shrotcomings like the creation of new jet collection for the sole purpose of adding new variables to the collection (as opposed to embedding the variables to the existing collection) and handling pat::Jet objects instead of using a pointer to the object. The previous code was basically an inefficient reinvention of better solutions already employed in NanoAOD FW.

Produces identical output to the previous code, but the speedup is O(100).