Run 100k events per one CRAB job
This implies that a NanoAOD event is produced in 500ms on average. This would mean that a CRAB job would finish in less than 14h. Although the absolute maximum limit is 21h and 55min, using slightly lower number of events per CRAB job leaves a bit room for the CRAB jobs finish on time in case the processing is a bit slower for some reason.