Keep AK8LS gen subjets if their pT > 8 GeV
I'm unable to find how the gen subjets in vanilla NanoAOD are filtered. There's a comment in NanoAOD saying that "in miniAOD only subjets from fat jets with pt > 100 are kept" but it's not clear to me how the cut on gen AK8 jet translates to cuts on AK8 gen subjets. I don't see any cuts on slimmedGenJetsAK8SoftDropSubJets, either. For these reasons, I'm setting the pT cut to that of AK4 gen jets in order to keep the increase in file size reasonable.
Somewhat related (possibly related to the JER uncertainties of SD mass in AK8LS collection): https://github.com/HEP-KBFI/tth-htt/issues/136